Fix #4038: Support >2 billion rigid bodies in fix rigid/small
Changed nbody from int to bigint and fixed MPI type mismatch.
Added unit tests for bigint type capacity and MPI communication.

(cherry picked from commit 07d7b0c29f2c5d20407760d2ad9465b29f9a55c2)

Author:    Michael Lostica <mlostica@gmail.com>
Date:      Sat Jul 25 11:19:09 2026 -0400

+++ src/RIGID/fix_rigid_small.h
@@ -78,7 +78,7 @@ class FixRigidSmall : public Fix {
   int earlyflag;       // 1 if forces/torques are computed at post_force()
   int commflag;        // various modes of forward/reverse comm
   int customflag;      // 1 if custom property/variable define bodies
-  int nbody;           // total # of rigid bodies
+  bigint nbody;        // total # of rigid bodies (supports >2^31)
   int nlinear;         // total # of linear rigid bodies
   tagint maxmol;       // max mol-ID
   double maxextent;    // furthest distance from body owner to body atom
